+++ src/components/index.js
@@ -4,3 +4,4 @@ export { default as Spinner } from './Spinner';
 export { default as Navigation } from './Navigation';
 export { default as GeneralError } from './GeneralError';
 export { default as GeneralAuth } from './GeneralAuth';
+export { default as AlertSnackbar } from './AlertSnackbar';